module5/runcontroller/requirements.txt
Flask==3.0.0
opentelemetry-distro==0.41b0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
redis==5.0.1
